Extract  app.js configurations, configured new watch for sass and js files, added hot reload.
