+++ package.json
@@ -81,6 +81,7 @@
   "devDependencies": {
     "@hmcts/eslint-config": "^1.4.0",
     "@types/chai": "^4.1.4",
+    "@types/chokidar": "^1.7.5",
     "@types/cli-color": "^0.3.29",
     "@types/config": "0.0.34",
     "@types/connect-redis": "0.0.7",
@@ -93,12 +94,14 @@
     "@types/node": "^9.4.6",
     "@types/nunjucks": "^3.1.0",
     "@types/puppeteer": "^1.6.4",
+    "@types/shelljs": "^0.8.3",
     "@types/sinon": "^5.0.2",
     "@types/sinon-chai": "^3.2.0",
     "browserify": "^16.2.3",
     "chai": "^3.5.0",
     "chai-as-promised": "^7.1.1",
     "chai-http": "^4.2.0",
+    "chokidar": "^2.1.1",
     "cli-color": "^1.3.0",
     "concurrently": "^3.6.1",
     "cross-env": "^4.0.0",
@@ -117,6 +120,8 @@
     "pa11y": "https://github.com/chrisgrimble/pa11y/tarball/v5.1.0-HMCTS-BUG-FIX",
     "pre-commit": "^1.2.2",
     "puppeteer": "^1.10.0",
+    "reload": "^2.4.0",
+    "shelljs": "^0.8.3",
     "sinon": "^2.1.0",
     "sinon-chai": "^2.9.0",
     "sonar-scanner": "^3.0.3",
